Implement RoomDao
