+++ functions/src/db/assertRightsToEditRoom.ts
@@ -1,13 +1,13 @@
 import { RoomId, Room } from '../types/Room'
 import { UID } from '../types/uid'
 import { ForbiddenError } from '../api/errors/HttpError'
-import { getRoom } from './getRoom'
+import { RoomDao } from './RoomDao'
 
 export const assertRightToEditRoom = async (
     roomId: RoomId,
     userId: UID
 ): Promise<Room> => {
-    const room = await getRoom(roomId)
+    const room = await RoomDao.get(roomId)
 
     if (room.uid !== userId) {
         throw new ForbiddenError('Not authorized to edit this resource')
